refactor: split flat tests/ into directory-based test architecture
- tests/unit/ — pure logic, mocks only, parallelizable via pytest-xdist
- tests/e2e/ — Zotero web API via MCP Client, auto-skip without API key
- tests/external/ — real external API calls (Crossref)
- tests/live/ — requires running local Zotero instance, auto-skip
- Per-directory conftest with pytest markers and auto-skip logic
- Add pytest-xdist dependency for parallel unit test execution
- Split test_doi.py, test_server_startup.py, test_live_instance.py
  across categories based on what they actually test
