+++ pyproject.toml
@@ -63,6 +63,12 @@ ignore_missing_imports = false
 asyncio_mode = "auto"
 testpaths = ["tests"]
 timeout = 60
+markers = [
+    "unit: pure logic, mocks only, safe to parallelize",
+    "e2e: Zotero web API via MCP Client",
+    "external: real external API calls (Crossref, etc.)",
+    "live: requires running local Zotero instance",
+]
 
 [tool.black]
 line-length = 100
@@ -79,4 +85,5 @@ dev = [
     "pre-commit>=3.8.0",
     "ipython>=8.0.0",
     "pytest-timeout>=2.4.0",
+    "pytest-xdist>=3.5.0",
 ]
