+++ tests/e2e/test_remove_item.py
