+++ tests/e2e/test_resources.py
@@ -3,7 +3,7 @@
 import pytest
 from fastmcp import Client
 
-from tests.test_helpers import ZoteroTestDataManager
+from tests.e2e.test_helpers import ZoteroTestDataManager
 from yazot.mcp_server import mcp
 from yazot.zotero_client import ZoteroClient
