+++ tests/unit/test_external_fulltext.py
